requirements.txt
flask==2.3.3
watchdog==3.0.0
deepgram-sdk>=3.0.0
openai>=1.3.0
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.0
tqdm==4.66.0